Refactored namespaces of JUS.Tool
