+++ src/JUS.Tool/Containers/Converters/BinaryFormat2Alar3.cs
@@ -1,11 +1,10 @@
 ﻿using System;
 using System.IO;
-using JUSToolkit.Formats.ALAR;
 using Yarhl.FileFormat;
 using Yarhl.FileSystem;
 using Yarhl.IO;
 
-namespace JUSToolkit.Converters.Alar
+namespace JUSToolkit.Containers.Converters
 {
     /// <summary>
     /// Converter between BinaryFormat and Alar3.
