+++ src/JUS.Tool/Graphics/Converters/Binary2Koma.cs
@@ -21,7 +21,7 @@ using Yarhl.IO;
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-namespace Texim.Games.JumpUltimateStars
+namespace JUSToolkit.Graphics.Converters
 {
     /// <summary>
     /// Converts between BinaryFormat and Koma.
