+++ src/JUS.Tool/Graphics/Koma.cs
@@ -20,7 +20,7 @@ using Yarhl.FileFormat;
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-namespace Texim.Games.JumpUltimateStars
+namespace JUSToolkit.Graphics
 {
     /// <summary>
     /// Sprite Format - A Koma is a Collection of KomaElements.
