+++ src/JUS.Tool/Graphics/KomaElement.cs
@@ -17,7 +17,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-namespace Texim.Games.JumpUltimateStars
+namespace JUSToolkit.Graphics
 {
     /// <summary>
     /// Image Format - Individual Koma OAM (square).
